+++ src/api/module-api.js
@@ -16,7 +16,7 @@ export function getModules() {
  */
 export function executeModule(name, options) {
   return axios.post(`/modules/${name}`, options)
-    .then(({ data }) => ({ agent: options.Agent, message: data.msg }))
+    .then(({ data }) => ({ agent: options.Agent, message: data.msg, taskID: data.taskID }))
     // eslint-disable-next-line prefer-promise-reject-errors
     .catch(error => Promise.reject({ agent: options.Agent, error: error.response.data.error }));
 }