+++ src/views/AgentEdit.vue
@@ -5,145 +5,81 @@
     >
       <v-breadcrumbs :items="breads" />
       <div class="pt-2">
-        <!-- TODO Wrap this into a component? -->
-        <v-dialog
-          ref="nameDialog"
-          v-model="dialog"
-          max-width="500px"
-        >
-          <v-card>
-            <v-card-title>
-              <span class="headline">Rename</span>
-            </v-card-title>
-            <v-card-text>
-              <v-form
-                ref="nameForm"
-              >
-                <v-container>
-                  <v-row>
-                    <v-col
-                      cols="12"
-                    >
-                      <v-text-field
-                        v-model="nameForm.name"
-                        label="Name"
-                        :rules="nameRules['name']"
-                        outlined
-                        dense
-                        required
-                      />
-                    </v-col>
-                  </v-row>
-                </v-container>
-              </v-form>
-            </v-card-text>
-            <v-card-actions>
-              <v-spacer />
-              <v-btn
-                color="blue darken-1"
-                text
-                @click="dialog = false"
-              >
-                Close
-              </v-btn>
-              <v-btn
-                color="blue darken-1"
-                text
-                :loading="nameLoading"
-                @click="renameAgent"
-              >
-                Save
-              </v-btn>
-            </v-card-actions>
-          </v-card>
-        </v-dialog>
-        <v-tooltip
-          bottom
-        >
-          <template v-slot:activator="{ on }">
-            <v-btn
-              color="primary"
-              class="mr-2"
-              fab
-              x-small
-              v-on="on"
-              @click="dialog = true"
-            >
-              <v-icon
-                style="padding-left: 4px"
-              >
-                fa-user-edit
-              </v-icon>
-            </v-btn>
-          </template>
-          <span>Rename Agent</span>
-        </v-tooltip>
-        <v-tooltip bottom>
-          <template v-slot:activator="{ on }">
-            <v-btn
-              color="primary"
-              class="mr-2"
-              fab
-              x-small
-              v-on="on"
-              @click="clearQueue"
-            >
-              <v-icon>fa-calendar-times</v-icon>
-            </v-btn>
-          </template>
-          <span>Clear Queued Tasks</span>
-        </v-tooltip>
-        <v-tooltip bottom>
-          <template v-slot:activator="{ on }">
-            <v-btn
-              color="error"
-              class="mr-2"
-              fab
-              x-small
-              v-on="on"
-              @click="killAgent"
-            >
-              <v-icon>fa-trash-alt</v-icon>
-            </v-btn>
-          </template>
-          <span>Kill Agent</span>
-        </v-tooltip>
-        <v-tooltip
-          v-if="!isChild"
-          bottom
-        >
-          <template v-slot:activator="{ on }">
-            <v-btn
-              fab
-              x-small
-              v-on="on"
-              @click="popout"
-            >
-              <v-icon>fa-external-link-alt</v-icon>
-            </v-btn>
-          </template>
-          <span>New Window</span>
-        </v-tooltip>
+        <agent-name-dialog
+          v-model="nameDialog"
+          :loading="nameLoading"
+          @submit="renameAgent"
+        />
+        <agent-upload-dialog
+          v-model="uploadDialog"
+          :language="agent.language"
+          :loading="uploadLoading"
+          @submit="doUpload"
+        />
+        <agent-download-dialog
+          v-model="downloadDialog"
+          :loading="downloadLoading"
+          @submit="doDownload"
+        />
+        <agent-tooltip-button
+          icon="fa-user-edit"
+          text="Rename Agent"
+          color="primary"
+          :pad-left="4"
+          @click="nameDialog = true"
+        />
+        <agent-tooltip-button
+          icon="fa-calendar-times"
+          text="Clear Queued Tasks"
+          color="primary"
+          @click="clearQueue"
+        />
+        <agent-tooltip-button
+          icon="fa-upload"
+          text="Upload"
+          @click="uploadDialog = true"
+        />
+        <agent-tooltip-button
+          icon="fa-download"
+          text="Download"
+          @click="downloadDialog = true"
+        />
+        <agent-tooltip-button
+          icon="fa-external-link-alt"
+          text="New Window"
+          @click="popout"
+        />
+        <agent-tooltip-button
+          icon="fa-trash-alt"
+          text="Kill Agent"
+          color="error"
+          @click="killAgent"
+        />
       </div>
     </div>
     <div :style="splitPaneHeight()">
       <split-pane
         :min-percent="20"
         :default-percent="60"
-        split="horizontal"
+        split="vertical"
       >
         <template slot="paneL">
           <v-tabs
             v-model="activeTab"
             class="scrollable-pane"
-            centered
-            small
+            fixed-tabs
           >
             <v-tab
               key="interact"
               href="#tab-interact"
             >
               Interact
+            </v-tab>
+            <v-tab
+              key="tasks"
+              href="#tab-tasks"
+            >
+              Tasks
             </v-tab>
             <v-tab
               key="view"
@@ -168,6 +104,20 @@
               </v-card>
             </v-tab-item>
             <v-tab-item
+              key="tasks"
+              :value="'tab-tasks'"
+            >
+              <v-card
+                class="scrollable-pane"
+                flat
+              >
+                <agent-command-history
+                  :agent-name="agent.name"
+                  :task-results="taskResults"
+                />
+              </v-card>
+            </v-tab-item>
+            <v-tab-item
               key="view"
               :value="'tab-view'"
             >
@@ -183,10 +133,12 @@
         <template slot="paneR">
           <div
             ref="bottomScrollable"
-            class="bottom-pane"
+            class="right-pane"
           >
             <agent-command-viewer
               :name="this.$route.params.id"
+              :task-results="taskResults"
+              :initialized="initialized"
               @new-results="scrollResults"
             />
           </div>
@@ -199,8 +151,13 @@
 <script>
 import AgentForm from '@/components/agents/AgentForm.vue';
 import AgentInteract from '@/components/agents/AgentInteract.vue';
+import AgentCommandHistory from '@/components/agents/AgentCommandHistory.vue';
 import AgentExecuteModule from '@/components/agents/AgentExecuteModule.vue';
 import AgentCommandViewer from '@/components/agents/AgentCommandViewer.vue';
+import AgentNameDialog from '@/components/agents/AgentNameDialog.vue';
+import AgentUploadDialog from '@/components/agents/AgentUploadDialog.vue';
+import AgentDownloadDialog from '@/components/agents/AgentDownloadDialog.vue';
+import AgentTooltipButton from '@/components/agents/AgentTooltipButton.vue';
 import SplitPane from 'vue-splitpane';
 import * as agentApi from '@/api/agent-api';
 
@@ -214,21 +171,26 @@ export default {
     AgentInteract,
     AgentExecuteModule,
     AgentCommandViewer,
+    AgentCommandHistory,
+    AgentNameDialog,
+    AgentUploadDialog,
+    AgentDownloadDialog,
+    AgentTooltipButton,
     SplitPane,
   },
   data() {
     return {
       agent: {},
-      nameRules: {
-        name: [
-          v => !!v || 'Name is required',
-          v => (!!v && v.length > 4) || 'Name must be at least 5 characters',
-        ],
-      },
       nameLoading: false,
-      nameForm: {},
+      uploadLoading: false,
+      downloadLoading: false,
       activeTab: 'View',
-      dialog: false,
+      nameDialog: false,
+      uploadDialog: false,
+      downloadDialog: false,
+      interval: null,
+      taskResults: [],
+      initialized: false,
     };
   },
   computed: {
@@ -264,6 +226,9 @@ export default {
   mounted() {
     this.getAgent(this.$route.params.id);
   },
+  beforeDestroy() {
+    clearInterval(this.interval);
+  },
   methods: {
     splitPaneHeight() {
       /* Not the prettiest thing, but seems to cover most window sizes to avoid page scroll.
@@ -278,6 +243,14 @@ export default {
       agentApi.getAgent(id)
         .then((data) => {
           this.agent = data;
+          if (this.interval !== null) {
+            clearInterval(this.interval);
+            this.interval = null;
+          }
+          this.interval = setInterval(async () => {
+            this.taskResults = await agentApi.getResults(this.agent.name);
+            this.initialized = true;
+          }, 5000);
         });
     },
     async killAgent() {
@@ -293,12 +266,12 @@ export default {
         this.$toast.success(`Clearing queued tasks for Agent ${this.agent.name}.`);
       }
     },
-    async renameAgent() {
-      if (this.nameLoading || !this.$refs.nameForm.validate()) { return; }
+    async renameAgent({ name }) {
+      if (this.nameLoading || name == null || name.length < 5) { return; }
 
       this.nameLoading = true;
       try {
-        await this.$store.dispatch('agent/rename', { oldName: this.agent.name, newName: this.nameForm.name });
+        await this.$store.dispatch('agent/rename', { oldName: this.agent.name, newName: name });
         this.$toast.success(`Agent ${this.agent.name} tasked to change name.`);
         this.$router.push({ name: 'agents' });
       } catch (err) {
@@ -306,7 +279,33 @@ export default {
       }
 
       this.nameLoading = false;
-      this.dialog = false;
+      this.nameDialog = false;
+    },
+    async doUpload({ file, pathToFile }) {
+      if (this.uploadLoading || file == null || pathToFile == null || pathToFile.length < 1) return;
+
+      this.uploadLoading = true;
+      try {
+        await agentApi.uploadFile(this.agent.name, file, pathToFile);
+      } catch (err) {
+        this.$toast.error(`Error: ${err}`);
+      }
+
+      this.uploadLoading = false;
+      this.uploadDialog = false;
+    },
+    async doDownload({ pathToFile }) {
+      if (this.downloadLoading || pathToFile == null || pathToFile.length < 1) return;
+
+      this.downloadLoading = true;
+      try {
+        await agentApi.downloadFile(this.agent.name, pathToFile);
+      } catch (err) {
+        this.$toast.error(`Error: ${err}`);
+      }
+
+      this.downloadLoading = false;
+      this.downloadDialog = false;
     },
     /**
      * When new results appear in the AgentCommandViewer, scroll it down to the bottom.
@@ -324,7 +323,7 @@ export default {
   overflow: auto;
 }
 
-.bottom-pane {
+.right-pane {
   background-color: white;
   height: 100%;
   overflow-y: auto;