+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] - 2021-01-18
+- Added a plugin list page and a plugin execution page
+- Added upload/download functionality to the agent page
+- Converted agents page to be left-right instead of top-bottom
+- Lots of code refactoring on agents
+
 ## [1.5.1] - 2020-12-26
 - Removed socket notifications from popout windows.
 - Minor dependency changes.